refactor: 모달 컨트롤러 내부 의존성을 메인 컨트롤러로 분리
